tests/wpt/mozilla/tests/mozilla/microdata/none_check.html

<!doctype html>
<meta charset="utf-8">
<html>
<head>
  <script src="../../../../web-platform-tests/resources/testharness.js"></script>
  <script src="../../../../web-platform/testharnessreport.js"></script>
  <title>None check test</title>
</head>

<body id="body">

  <h1 id="h1">Header H1</h1>
  <h2 id="h2">Header H2</h2>

  <p id="p">Paragraph</p>

  <form id="form">
    <input id="input">Input Field in form</input>
  </form>

  <ul id="ul">
    <li id="li">Unordered List Item</li>
  </ul>

  <meta id="meta">Meta Tag</meta>

  <table id="table">Table</table>

  <div id="div">Hi</div>

  <a id="anchor"></a>

</body>

<script>
  test(function(){
    // Test body element
    var body = document.getElementById("body");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test h1 element
    var h1 = document.getElementById("h1");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test h2 element
    var h2 = document.getElementById("h2");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test p element
    var p = document.getElementById("p");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test input element
    var input = document.getElementById("input");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test form element
    var form = document.getElementById("form");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test li element
    var li = document.getElementById("li");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test ul element
    var ul = document.getElementById("ul");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test meta element
    var meta = document.getElementById("meta");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test table element
    var table = document.getElementById("table");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test div element
    var div = document.getElementById("div");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

    // Test anchor element
    var anchor = document.getElementById("anchor");
    assert_equals(body.propertyNames(), null, "The method should return null ");
    assert_equals(body.itemtypes(), null, "The method should return null ");

  });
</script>

</html>
